+++ config/business/upload/const.rs
@@ -4,4 +4,4 @@ pub const STATIC_ROUTE: &'static str = "static";
 pub const UPLOAD_DIR_KEY: &'static str = "upload_dir";
 pub const UPLOAD_FILE_KEY: &'static str = "upload_file";
 pub const UPLOAD_HTML: &'static str = include_str!("../../../resources/static/upload/index.html");
-pub const PARTIAL_CONTENT_STATUS: u16 = 206;
+pub const PARTIAL_CONTENT_STATUS: usize = 206;
